--- refer/__init__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-refer
-=====
-This library provides several utillity functions to load ReferIt split
-instances.
-"""
-
-from .refer import REFER
-
-REFER
-__author__ = 'licheng'
-
-VERSION_INFO = (1, 0, 0, 'dev0')
-__version__ = '.'.join(map(str, VERSION_INFO))
