--- refer/evaluation/bleu/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'tylin'
