--- refer/evaluation/rouge/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'vrama91'
