--- refer/external/.gitignore
@@ -1,3 +1,0 @@
-*.pyc
-_mask.so
-_mask.c
