--- refer/external/README.md
@@ -1,1 +1,0 @@
-The codes inside this folder are copied from pycocotools: https://github.com/pdollar/coco
