+++ referit/external/README.md
@@ -1,0 +1,1 @@
+The codes inside this folder are copied from pycocotools: https://github.com/pdollar/coco
